fix feature-redux (AppコンポーネントにModal要素移動・Modalの処理をreduxに変更)
